Factor out update full and incremental partner classes
